backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
httpx==0.28.1
requests==2.33.1
anthropic==0.89.0
